+++ app/src/main/java/mustafaozhan/github/com/androcat/notification/NotificationReceiver.kt
@@ -1,21 +1,11 @@
 package mustafaozhan.github.com.androcat.notification
 
 import android.app.AlarmManager
-import android.app.NotificationChannel
-import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
-import android.graphics.BitmapFactory
-import android.os.Build
 import android.util.Log
-import androidx.annotation.RequiresApi
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-import mustafaozhan.github.com.androcat.R
 import mustafaozhan.github.com.androcat.base.BaseBroadcastReceiver
-import mustafaozhan.github.com.androcat.notification.model.Notification
-import mustafaozhan.github.com.androcat.notification.model.NotificationType
 
 class NotificationReceiver : BaseBroadcastReceiver() {
     companion object {
@@ -35,7 +25,7 @@ class NotificationReceiver : BaseBroadcastReceiver() {
                 .subscribe(
                     {
                         it.forEach { notification ->
-                            senNotification(notification, context)
+                            NotificationUtil.senNotification(notification, context)
                         }
                         compositeDisposable.dispose()
                     }, {
@@ -77,43 +67,4 @@ class NotificationReceiver : BaseBroadcastReceiver() {
         alarmManager?.cancel(pendingIntent)
     }
 
-    private fun senNotification(notification: Notification, context: Context) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            createNotificationChannel(context)
-        }
-
-        with(NotificationManagerCompat.from(context)) {
-            notify(
-                notification.id?.toInt() ?: -1,
-                NotificationCompat
-                    .Builder(context, NotificationCompat.CATEGORY_SOCIAL)
-                    .setSmallIcon(
-                        when (notification.subject?.type.toString()) {
-                            NotificationType.ISSUE.value -> R.drawable.ic_issue
-                            NotificationType.PULL_REQUEST.value -> R.drawable.ic_pull_request
-                            else -> R.drawable.ic_androcat_dash
-                        }
-                    )
-                    .setLargeIcon(BitmapFactory.decodeResource(context.resources, R.mipmap.ic_launcher_round))
-                    .setContentTitle("[${notification.subject?.type}] ${notification.repository?.fullName}")
-                    .setContentText(notification.subject?.title.toString())
-                    .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-                    .build()
-            )
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun createNotificationChannel(context: Context) {
-
-        val channel = NotificationChannel(
-            NotificationCompat.CATEGORY_SOCIAL,
-            context.getString(R.string.app_name),
-            NotificationManager.IMPORTANCE_DEFAULT
-        )
-
-        val notificationManager: NotificationManager? =
-            context.getSystemService(Context.NOTIFICATION_SERVICE) as? NotificationManager
-        notificationManager?.createNotificationChannel(channel)
-    }
 }
